Used dir for models and views instead of py files
