+++ docs/conf.py
@@ -12,8 +12,11 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('..'))
 
+from django.conf import settings
+
+sys.path.insert(0, os.path.abspath('..'))
+settings.configure()
 
 # -- Project information -----------------------------------------------------
 
@@ -35,12 +38,10 @@ extensions = [
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
-
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
-
 
 # -- Options for HTML output -------------------------------------------------
