+++ docs/record_trades.rst
@@ -1,0 +1,21 @@
+record\_trades package
+=======================
+
+Submodules
+----------
+
+record_trades module
+-------------------------------
+
+.. autoclass:: record_trades.models.trade
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: record_trades
+   :members:
+   :undoc-members:
+   :show-inheritance:
