--- docs/trade_recorder.rst
@@ -1,45 +1,0 @@
-trade\_recorder package
-=======================
-
-Submodules
-----------
-
-trade\_recorder.asgi module
----------------------------
-
-.. automodule:: trade_recorder.asgi
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-trade\_recorder.settings module
--------------------------------
-
-.. automodule:: trade_recorder.settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-trade\_recorder.urls module
----------------------------
-
-.. automodule:: trade_recorder.urls
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-trade\_recorder.wsgi module
----------------------------
-
-.. automodule:: trade_recorder.wsgi
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: trade_recorder
-   :members:
-   :undoc-members:
-   :show-inheritance:
