+++ record_trades/models/__init__.py
@@ -1,0 +1,4 @@
+"""
+"""
+
+from .trade import Trade
